refactor: rename DocumentRepository to DocumentsRepository and update related classes
