--- app/src/main/kotlin/me/tomasan7/jecnamobile/document/DocumentsState.kt
@@ -1,5 +1,0 @@
-package me.tomasan7.jecnamobile.document
-
-class DocumentsState
-{
-}
