+++ app/src/main/kotlin/me/tomasan7/jecnamobile/documents/DocumentsRepositoryImpl.kt
@@ -1,12 +1,12 @@
-package me.tomasan7.jecnamobile.document
+package me.tomasan7.jecnamobile.documents
 
 import io.github.tomhula.jecnaapi.JecnaClient
 import io.github.tomhula.jecnaapi.data.document.DocumentsPage
 import javax.inject.Inject
 
-class DocumentRepositoryImpl @Inject constructor(
+class DocumentsRepositoryImpl @Inject constructor(
     private val client: JecnaClient
-) : DocumentRepository
+) : DocumentsRepository
 {
     override suspend fun getDocumentsPage(): DocumentsPage = client.getDocumentsPage()
 }
